+++ dap/host.go
@@ -8,23 +8,24 @@ import (
 
 // eventHost receives events from the debugger
 type eventHost struct {
-	handler *Session
-	noDebug bool
+	// send function is concurrency safe
+	send func(api.Message)
+
+	// copy variables from session to avoid memory races
+	source  api.Source
+	lineOff int
+	colOff  int
 }
 
 func (eh eventHost) Paused(reason string) {
-	eh.handler.send(&api.StoppedEvent{
+	eh.send(&api.StoppedEvent{
 		Event: *newEvent("stopped"),
 		Body:  api.StoppedEventBody{Reason: reason, ThreadId: 1, AllThreadsStopped: true},
 	})
 }
 
-func (eh eventHost) BreakOnException() bool {
-	return !eh.noDebug
-}
-
 func (eh eventHost) Exception(err error) {
-	eh.handler.send(&api.StoppedEvent{
+	eh.send(&api.StoppedEvent{
 		Event: *newEvent("stopped"),
 		Body:  api.StoppedEventBody{Reason: "exception", Description: "exception", Text: err.Error(), ThreadId: 1, AllThreadsStopped: true},
 	})
@@ -32,23 +33,23 @@ func (eh eventHost) Exception(err error) {
 
 func (eh eventHost) Panicked(err error, pos []ast.Position, trace []string) {
 	// just send stack traces over stderr and exit
-	eh.handler.send(&api.OutputEvent{
+	eh.send(&api.OutputEvent{
 		Event: *newEvent("output"),
 		Body: api.OutputEventBody{
 			Category: "stderr",
 			Output:   "error: " + err.Error() + "\n",
-			Source:   &eh.handler.source,
+			Source:   &eh.source,
 		},
 	})
 	for i := range pos {
-		eh.handler.send(&api.OutputEvent{
+		eh.send(&api.OutputEvent{
 			Event: *newEvent("output"),
 			Body: api.OutputEventBody{
 				Category: "stderr",
 				Output:   "\tin " + trace[i] + "\n",
-				Source:   &eh.handler.source,
-				Line:     pos[i].Line + eh.handler.lineOff,
-				Column:   pos[i].Column + eh.handler.colOff,
+				Source:   &eh.source,
+				Line:     pos[i].Line + eh.lineOff,
+				Column:   pos[i].Column + eh.colOff,
 			},
 		})
 	}
@@ -56,7 +57,7 @@ func (eh eventHost) Panicked(err error, pos []ast.Position, trace []string) {
 }
 
 func (eh eventHost) Exited(code int) {
-	eh.handler.send(&api.ExitedEvent{
+	eh.send(&api.ExitedEvent{
 		Event: *newEvent("exited"),
 		Body:  api.ExitedEventBody{ExitCode: code},
 	})
@@ -64,7 +65,7 @@ func (eh eventHost) Exited(code int) {
 }
 
 func (eh eventHost) Terminated() {
-	eh.handler.send(&api.TerminatedEvent{
+	eh.send(&api.TerminatedEvent{
 		Event: *newEvent("terminated"),
 	})
 }
